move non-streamlit github utils to dcpy
