+++ apps/qaqc/tests/test_github.py
@@ -1,8 +1,9 @@
-from src.github import get_default_branch, get_branches
+from dcpy.git.github import get_default_branch, get_branches
 
 TEST_REPO = "data-engineering-qaqc"
 
 
+## TODO - these tests don't belong here since this functionality has been moved to dcpy
 def test_get_default_branch():
     branch = get_default_branch(repo=TEST_REPO)
     assert branch == "main"
