+++ dcpy/git/git.py
@@ -1,7 +1,7 @@
 import os
 import subprocess
 
-from . import REPO_ROOT_PATH
+from .. import REPO_ROOT_PATH
 
 
 def git_username():
